requirements.txt
requests==2.24.0
netifaces==0.10.9
python-dotenv==0.14.0